+++ src/mod.ts
@@ -1,14 +1,18 @@
-export type * from "./coreTypes/mod";
 export * from "./limits";
+export * from "./Toodle";
+
+export * as Colors from "./colors/mod";
 export * as GfxMath from "./math/mod";
 export * as Scene from "./scene/mod";
-export type * from "./scene/QuadNode";
-export type * from "./scene/SceneNode";
 export * as Screen from "./screen/mod";
 export * as Shaders from "./shaders/mod";
-export type * from "./shaders/QuadShader";
-export * from "./Toodle";
 export * as Text from "./text/mod";
-export type * from "./text/TextNode";
 export * as Textures from "./textures/mod";
 export * as Utils from "./utils/mod";
+
+export type * from "./coreTypes/mod";
+
+export type * from "./scene/QuadNode";
+export type * from "./scene/SceneNode";
+export type * from "./shaders/QuadShader";
+export type * from "./text/TextNode";