+++ src/shaders/mod.ts
@@ -1,4 +1,5 @@
 export * from "./EngineUniform";
 export * from "./IShader";
+export * from "./postprocess/mod";
 export * from "./ShaderDescriptor";
 export * from "./samplers";