+++ docs/CHANGELOG.md
@@ -4,7 +4,16 @@ A detailed list of all notable changes for every released version.
 
 [All releases](https://www.npmjs.com/package/@bloop.gg/toodle)
 
-## [0.0.96](https://github.com/bloopgames/toodle/releases/tag/0.0.95)
+## [0.0.99](https://github.com/bloopgames/toodle/releases/tag/0.0.97)
+
+- Add low level api for post-process (fullscreen) shaders, see [post-process example](https://toodle.gg/examples/post-process.html)
+- Add `Colors.web` for CSS web color names to use when prototyping
+
+## [0.0.97](https://github.com/bloopgames/toodle/releases/tag/0.0.97)
+
+- Add MIT license to package.json so it shows up on npm and gh.
+
+## [0.0.96](https://github.com/bloopgames/toodle/releases/tag/0.0.96)
 
 - Fix: `toodle.endFrame` now clears instance counts even if an uncaught error is thrown to avoid spamming the console with a ToodleInstanceCap error